Implement agent signals (#91)
Agents can now define `signals-from` at agent level. The value must be a
string and a reference to a defined topic.

Each agent will automatically have an hook and a consumer to that topic.
State storage and wecbcrawler implements it by hadling logic to
invalidate partially or fully the state.


For s3/azure/gcs:
```
type: s3-source
signals-from: my-signals-topic
```  
based on the key the will behave in this way:
- `key==invalidate`. Will check the value as string and invalidate that
object state, if exists.
- `key==invalidate-all`. Will invalidate the entire state

For webcrawler, only `key==invalidate-all` is implemented. It makes the
state to be reset and a new indexing triggered.