gateways: allow authorization via http header (#93)
Added flag to the gateway inside `authentication` named
`http-credentials-source` defaults to `query`. Possible values are
`query` and `header`.

If using an http endpoint, the api gateway will check this gateway flag
to gather the credentials. This apply both to `produce` and `service`
endpoints.

CLI has been updated to decide the auth mode based on the gateway config
(only for produce with --protocol http)